backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
pytest>=8.0.0
httpx>=0.27.0
psutil>=5.9.0
